separating the helper out into its own header. Updated the test.
